高雅人士2

+++ include/core/App.hpp
@@ -23,9 +23,9 @@ class App
     static constexpr int bgheight = 896;
     static constexpr int bgoffsetX = 64;
     static constexpr int bgoffsetY = 32;
-    static constexpr glm::vec2 bgCenter = {bgoffsetX + 0.5 * bgwidth, bgoffsetY + 0.5 * bgheight};
+    static constexpr glm::vec2 bgCenter = {bgoffsetX + 0.5f * bgwidth, bgoffsetY + 0.5f * bgheight};
     inline static int scale = 1;
-    static constexpr double STAT_INTERVAL = 1.0f;
+    static constexpr float STAT_INTERVAL = 1.0f;
     static constexpr int ESTIMATE_BULLET = 500;
 
   private:
